uk-pmrv-app-web/src/app/permit-application/permit-type/permit-type-summary/permit-type-summary.guard.ts
import { Injectable } from '@angular/core';
import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';

import { map, Observable } from 'rxjs';

import { PermitApplicationStore } from '../../store/permit-application.store';

@Injectable({
  providedIn: 'root',
})
export class PermitTypeSummaryGuard implements CanActivate {
  constructor(private readonly store: PermitApplicationStore, private readonly router: Router) {}

  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean | UrlTree> {
    return this.store.pipe(
      map((storeState) => {
        return (
          !!storeState.permitType ||
          !storeState.isEditable ||
          this.router.parseUrl(state.url.slice(0, state.url.lastIndexOf('/summary')))
        );
      }),
    );
  }
}
